+++ NEWS.md
@@ -35,7 +35,7 @@
 ### Changed
 - The fields that had as type a `Matrix of size Px1` have now the type `Vector` 
 - Many new fields of the type `VSMSolution` documented
-- `init!(body_aero)` is now a public function
+- `reinit!(body_aero)` is now a public function
 ### Added
 - Added the option to use nonlinear solve to calculate the gamma distribution #140
 - New page `Tips and tricks` added to the documentation